paths issue fixed and login_screen.dart restructured
